+++ spec/unit/type/pkg_publisher_spec.rb
@@ -2,201 +2,215 @@ require 'spec_helper'
 
 describe Puppet::Type.type(:pkg_publisher) do
 
-  before do
-    @class = described_class
-    @profile_name = "rspec profile"
+  let(:params) do
+    {
+      :name => 'mpyub',
+      :origin => %w[http://pkg.oracle.com/solaris/release],
+      :ensure => :present,
+      :proxy => [],
+      :mirror => [],
+      :sslkey => [],
+      :sslcert => []
+    }
+  end
+  let(:resource) { described_class.new(params) }
+  let(:provider) { Puppet::Provider.new(resource) }
+  let(:catalog) { Puppet::Resource::Catalog.new }
+  let(:error_pattern) { /Invalid value/m }
+
+  it "has :name as its keyattribute" do
+    expect( described_class.key_attributes).to be == [:name]
   end
 
-  it "should have :name as its keyattribute" do
-    expect( @class.key_attributes).to be == [:name]
-  end
-
-  describe "when validating attributes" do
+  describe "has property" do
     [ :ensure, :origin, :enable, :sticky, :searchfirst, :searchafter,
       :searchbefore, :proxy, :sslkey, :sslcert, :mirror
     ].each do |prop|
-      it "should have a #{prop} property" do
-        expect(@class.attrtype(prop)).to be == :property
+      it prop do
+        expect(described_class.attrtype(prop)).to be == :property
       end
     end
   end
 
-  describe "when validating values" do
-
-    describe "for ensure" do
-      error_pattern = /Invalid value/m
-      def validate(ens)
-         @class.new(:name => @profile_name, :ensure => ens)
-      end
-
-      [ "present", "absent" ].each do |newval|
-        it "should accept a value of #{newval}" do
-          expect { validate(newval) }.not_to raise_error
+  describe "parameter" do
+    [:ensure].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(present absent).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
+        context "rejects" do
+          %w(online offline).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
         end
       end
-
-      it "should reject invalid values" do
-        expect { validate "foo" }.to raise_error(Puppet::Error, error_pattern)
-      end
-    end  # ensure
-
-    describe "for origin" do
-      def validate(orig)
-         @class.new(:name => @profile_name, :origin => orig)
-      end
-
-      it "should accept a value" do
-        expect { validate "[ 'file://mydir/' ]"
-             }.not_to raise_error
-      end
-
-      it "should return an array for a single value" do
-        mytype = @class.new(:name => @profile_name,
-                            :origin => "file://mydir")
-        expect(mytype.property("origin").value).to be_an(Array)
-      end
-
-      it "should trim trailing slashes from file:// URI values" do
-        mytype = @class.new(:name => @profile_name,
-                            :origin => ['file://trail/', 'file://notrail'])
-        expect(mytype.property("origin").value).to be ==
-                                       ['file://trail', 'file://notrail']
-      end
-    end  # origin
-
-    describe "for enable" do
-      error_pattern = /enable.*Invalid/m
-
-      def validate(enab)
-         @class.new(:name => @profile_name, :enable => enab)
-      end
-
-      [ "true", "false" ].each do |follow_val|
-        it "should accept a value of #{follow_val}" do
-          expect { validate(follow_val) }.not_to raise_error
+    end
+    [:origin, :mirror].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(file://mydir/ https://foo/bar http://foo/bar).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
+        context "rejects" do
+          %w(bare_string).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
+        end
+        it "returns an array for a single value" do
+          params[:origin] = "file://mydir"
+          expect(resource[:origin]).to eq([params[:origin]])
+        end
+        it "trims trailing slashes" do
+          params[type] = %w(file://trail/ file://notrail)
+          expect(resource[type]).to eq(%w(file://trail file://notrail))
         end
       end
-
-      it "should reject an invalid value" do
-        expect { validate("foobar") }.to raise_error(Puppet::ResourceError,
-                                                       error_pattern)
-      end
-    end  # enable
-
-    describe "for sticky" do
-      error_pattern = /sticky.*Invalid/m
-
-      def validate(stickyval)
-         @class.new(:name => @profile_name, :sticky => stickyval)
-      end
-
-      [ "true", "false" ].each do |follow_val|
-        it "should accept a value of #{follow_val}" do
-          expect { validate(follow_val) }.not_to raise_error
+    end
+    [:enable, :sticky].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(true false) + [true, false].each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
+        context "rejects" do
+          %w(yes no).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
         end
       end
-
-      it "should reject an invalid value" do
-        expect { validate("foobar") }.to raise_error(Puppet::ResourceError,
-                                                       error_pattern)
-      end
-    end  # sticky
-
-    describe "for searchfirst" do
-      error_pattern = /searchfirst.*Invalid/m
-
-      def validate(sf)
-         @class.new(:name => @profile_name, :searchfirst => sf)
-      end
-
-      [ "true" ].each do |follow_val|
-        it "should accept a value of #{follow_val}" do
-          expect { validate(follow_val) }.not_to raise_error
+    end
+    [:searchfirst].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(true).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
+        context "rejects" do
+          %w(false).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
         end
       end
-
-      it "should reject an invalid value" do
-        expect { validate("foobar") }.to raise_error(Puppet::ResourceError,
-                                                       error_pattern)
+    end
+    [:searchafter, :searchbefore].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(a_string).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
       end
-    end  # searchfirst
-
-    describe "for searchafter" do
-      def validate(sa)
-         @class.new(:name => @profile_name, :searchafter => sa)
+    end
+    [:proxy].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(http://foo).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+            it "<undef>" do
+              params.delete(type)
+              expect { resource }.not_to raise_error
+            end
+        end
+        context "rejects" do
+          %w(https://bar).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
+        end
       end
-
-      it "should accept a value" do
-        expect { validate "foo" }.not_to raise_error
+    end
+    [:sslcert, :sslkey].each do |type|
+      context "#{type}" do
+        context "accepts" do
+          %w(/fully/specified/path).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.not_to raise_error
+            end
+          end
+        end
+        context "rejects" do
+          %w(~/expanded/path relative/path).each do |thing|
+            it thing.inspect do
+              params[type] = thing
+              expect { resource }.to raise_error(Puppet::Error, error_pattern)
+            end
+          end
+        end
       end
-    end  # searchafter
-
-    describe "for searchbefore" do
-      def validate(sb)
-         @class.new(:name => @profile_name, :searchbefore => sb)
+    end
+  end # parameters
+  context "autorequire" do
+    before do
+      params[:sslkey] = "/foo"
+      params[:sslcert] = "/bar"
+    end
+    context "files" do
+      it "does not require file when no matching resource exists" do
+        file = Puppet::Type.type(:file).new(:name => "/baz")
+        catalog.add_resource resource
+        catalog.add_resource file
+        expect(resource.autorequire.count).to eq 0
       end
-
-      it "should accept a value" do
-        expect { validate "foo" }.not_to raise_error
+      it "requires the matching file resource" do
+        sslkey = Puppet::Type.type(:file).new(:name => params[:sslkey])
+        catalog.add_resource sslkey
+        sslcert = Puppet::Type.type(:file).new(:name => params[:sslcert])
+        catalog.add_resource sslcert
+        catalog.add_resource resource
+        reqs = resource.autorequire
+        expect(reqs.count).to eq 2
+        expect(reqs[0].source).to eq sslkey
+        expect(reqs[0].target).to eq resource
+        expect(reqs[1].source).to eq sslcert
+        expect(reqs[1].target).to eq resource
       end
-    end  # searchbefore
-
-    describe "for proxy" do
-      def validate(pval)
-         @class.new(:name => @profile_name, :proxy => pval,
-         :origin => "http://bar")
+      it "does not fail if sslcert and key are nil" do
+        params.delete(:sslkey)
+        params.delete(:sslcert)
+        file = Puppet::Type.type(:file).new(:name => "/baz")
+        catalog.add_resource resource
+        catalog.add_resource file
+        expect(resource.autorequire.count).to eq 0
       end
-
-      it "should accept a value" do
-        expect { validate "http://foo" }.not_to raise_error
-      end
-    end  # proxy
-
-    describe "for sslkey" do
-      def validate(skey)
-         @class.new(:name => @profile_name, :sslkey => skey,
-         :origin => "http://bar")
-      end
-
-      it "should accept a value" do
-        expect { validate "/foo" }.not_to raise_error
-      end
-    end  # sslkey
-
-    describe "for sslcert" do
-      def validate(scert)
-         @class.new(:name => @profile_name, :sslcert => scert,
-         :origin => "http://bar")
-      end
-
-      it "should accept a value" do
-        expect { validate "/foo" }.not_to raise_error
-      end
-    end  # sslcert
-
-    describe "for mirror" do
-      def validate(mir)
-         @class.new(:name => @profile_name, :mirror => mir)
-      end
-
-      it "should accept a value" do
-        expect { validate "[ 'file://mydir/' ]"
-             }.not_to raise_error
-      end
-
-      it "should return an array for a single value" do
-        mytype = @class.new(:name => @profile_name,
-                            :mirror => "file://mydir")
-        expect(mytype.property("mirror").value).to be_an(Array)
-      end
-
-      it "should trim trailing slashes from file:// URI values" do
-        mytype = @class.new(:name => @profile_name,
-                            :mirror => ['file://trail/', 'file://notrail'])
-        expect(mytype.property("mirror").value).to be ==
-                                       ['file://trail', 'file://notrail']
-      end
-    end  # mirror
-
-  end # validating values
+    end
+  end
 end